requirements.txt
google-auth
google-cloud-bigquery
pandas
pandas-gbq
astronomer-cosmos
apache-airflow-providers-google
